Move cli int a sub-module
