+++ hyper_bump_it/_cli/to.py
@@ -1,0 +1,56 @@
+"""
+Bump to version command.
+"""
+from pathlib import Path
+from typing import Optional
+
+import typer
+
+from hyper_bump_it import _core as core
+from hyper_bump_it._cli import common
+from hyper_bump_it._config import BumpToArgs, GitAction, config_for_bump_to
+
+
+def to(
+    new_version: str = typer.Argument(
+        ..., help="The new version to bump to", show_default=False
+    ),
+    config_file: Optional[Path] = common.CONFIG_FILE,
+    project_root: Path = common.PROJECT_ROOT,
+    dry_run: bool = common.DRY_RUN,
+    skip_confirm_prompt: Optional[bool] = common.SKIP_CONFIRM_PROMPT,
+    current_version: Optional[str] = common.CURRENT_VERSION,
+    commit: Optional[GitAction] = common.COMMIT,
+    branch: Optional[GitAction] = common.BRANCH,
+    tag: Optional[GitAction] = common.TAG,
+    remote: Optional[str] = common.REMOTE,
+    commit_format_pattern: Optional[str] = common.COMMIT_FORMAT_PATTERN,
+    branch_format_pattern: Optional[str] = common.BRANCH_FORMAT_PATTERN,
+    tag_format_pattern: Optional[str] = common.TAG_FORMAT_PATTERN,
+) -> None:
+    """
+    Bump the version to a specific version.
+    """
+    new_version_parsed = common.parse_version(new_version, "NEW_VERSION")
+    current_version_parsed = common.parse_version(current_version, "--current-version")
+
+    with common.handle_bump_errors():
+        app_config = config_for_bump_to(
+            BumpToArgs(
+                new_version=new_version_parsed,
+                config_file=config_file,
+                project_root=project_root,
+                dry_run=dry_run,
+                skip_confirm_prompt=skip_confirm_prompt,
+                current_version=current_version_parsed,
+                commit=commit,
+                branch=branch,
+                tag=tag,
+                remote=remote,
+                commit_format_pattern=commit_format_pattern,
+                branch_format_pattern=branch_format_pattern,
+                tag_format_pattern=tag_format_pattern,
+            )
+        )
+
+        core.do_bump(app_config)
